changed in Order class int to long in setDriverID and setClientId changed in ConverterFactory in convertToEntity (long) fieldValue to Long.parseLong(fieldValue.toString()) and in convertToModel add chek if (!(entry.getValue() == null))

+++ src/main/java/com/netcracker/project/study/model/order/Order.java
@@ -41,7 +41,7 @@ public class Order extends Model{
         return driverId;
     }
 
-    public void setDriverId(int driverId) {
+    public void setDriverId(long driverId) {
         this.driverId = driverId;
     }
 
@@ -49,7 +49,7 @@ public class Order extends Model{
         return clientId;
     }
 
-    public void setClientId(int clientId) {
+    public void setClientId(long clientId) {
         this.clientId = clientId;
     }
